Reestructurar código en arquitectura modular
- Crear carpeta config/ para configuración centralizada (URL API, reglas de validación)
- Crear carpeta controllers/ con módulos específicos:
  * variables.js - Selectores del DOM y variables globales
  * funcionesUI.js - Manejo de interfaz y modales
  * funcionesFormulario.js - Limpieza y carga de formularios
  * funcionesTareas.js - CRUD de tareas y renderizado
  * index.js - Controlador principal con event listeners
- Simplificar app.js a solo importaciones e inicialización
- Mantener lógica exacta del código original sin modificaciones
- Agregar documentación detallada con comentarios en todos los archivos
